+++ src/runez/file.py
@@ -218,6 +218,22 @@ def ini_to_dict(path, keep_empty=False, fatal=False, logger=False):
     return result
 
 
+def is_subfolder(path, root_folder):
+    """
+    Args:
+        path (str): Path to file or folder
+        root_folder (str): Path to folder to consider as root
+
+    Returns:
+        (bool): True if 'path' is a subfolder of 'root_folder'
+    """
+    if path and root_folder:
+        abs_path = os.path.abspath(path)
+        abs_root = os.path.abspath(root_folder)
+        prefix = os.path.commonprefix([abs_path, abs_root])
+        return prefix == abs_root
+
+
 def is_younger(path, age, default=False):
     """
     Args:
@@ -592,6 +608,12 @@ def _untar(source, destination, simplify):
     with TempFolder():
         extracted_source = to_path(source.name)
         with tarfile.open(source) as fh:
+            for member in fh.getmembers():
+                # See https://www.trellix.com/en-us/about/newsroom/stories/research/tarfile-exploiting-the-world.html
+                member_path = os.path.join(extracted_source, member.name)
+                if not is_subfolder(member_path, extracted_source):  # pragma: no cover, don't have an exploit sample handy
+                    raise Exception("Attempted Path Traversal in Tar File")
+
             fh.extractall(extracted_source)
 
         _move_extracted(extracted_source, destination, simplify)