+++ docs/history.rst
@@ -2,6 +2,20 @@
 History
 =======
 
+3.6.2 (2023-03-02)
+------------------
+
+* Support a simple optional cache for RestClient-s
+
+* CVE-2007-4559
+
+
+3.6.1 (2022-06-01)
+------------------
+
+* Avoid creating anything at import time (removed most globals)
+
+
 3.6.0 (2022-05-26)
 ------------------